# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a new account (pending admin approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "202":
          description: Account created, pending approval
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "204":
          description: Logged in
        "401":
          description: Invalid password
...
  /auth/google:
    post:
      operationId: googleAuth
      tags: [auth]
      summary: Authenticate with a Google ID token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthInput"
      responses:
        "204":
          description: Logged in
        "401":
          description: Invalid token or email not registered
        "403":
          description: Account pending approval
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Update the application password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password updated
        "400":
          description: Invalid input
        "401":
          description: Current password is incorrect
  /auth/extension-token:
    get:
      operationId: getExtensionToken
      tags: [auth]
      summary: Get the extension API token (generates one on first call)
      responses:
        "200":
          description: Token value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtensionToken"
        "401":
          description: Authentication required
  /auth/extension-token/regenerate:
    post:
      operationId: regenerateExtensionToken
      tags: [auth]
      summary: Rotate the extension API token (old token immediately invalidated)
      responses:
        "200":
          description: New token value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtensionToken"
        "401":
          description: Authentication required
...
components:
  schemas:
...
    ExtensionToken:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 1
    GoogleAuthInput:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 6
          maxLength: 200
...